requirements.txt
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
torch==2.0.0
torchvision==0.15.1
yfinance==0.2.12
